PWD should not be HOME.
When the current working directory is the same as HOME this might
cause problems when building .NET Core apps. The problem stems
from some nuget packages including C# source files. When
"dotnet build" is invoked in a directory which is also the HOME
folder, it will attempt to recursively build all source files
under it. That might fail for some source files in $HOME/.nuget.

See issue 28 for details. The fix is to set HOME to one level
up. That is the current working directory will not be outside the
recursive search path when "dotnet build" is being invoked.

Closes issue 28.