Make all DOTNET_* env vars, build env vars only (#39)
Previously, DOTNET_* environment variables were used by
the s2i assemble and run script, which broke deployments in
some cases.

We now only use the DOTNET_* environment variables in the
assemble script (build environment) and generate a s2i_run
script on-the-fly which will be independent of DOTNET_* 
environment variables. The output of the assemble script is published
to folder /opt/app-root/publish. This includes an s2i_run script suitable
to start the .NET Core application.