Add DOTNET_ASSEMBLY_NAME envvar
Used to select the assembly to run. This must NOT include the `.dll`
extension. Set this to the output assembly name specified in `project.json`
(`buildoptions/outputName`) or MSBuild (`PropertyGroup/AssemblyName`).
For `project.json`, the assembly name defaults to the `project.json` parent
folder. This value is used as the default value for `DOTNET_ASSEMBLY_NAME`.
When `project.json` is at the `context-dir`, the parent folder name will
be 'src'. So, by default, this generates a 'src.dll' assembly. Setting
`DOTNET_ASSEMBLY_NAME` will cause the project to be built in a folder
named `<DOTNET_ASSEMBLY_NAME>` which builds a `<DOTNET_ASSEMBLY_NAME>.dll`.