Add a simple set of tests for released images
The asp-net-hello-world application is used as a sample application by
OpenShift for released versions of this image. We need a smaller set of
tests to ensure that the released versions of this image continue
working with asp-net-hello-world application.

Unlike the existing tests, this test should just check that the image
'works'. It doesn't have to be exhaustive and shouldn't test all
applications, just the one that will be used by OpenShift. Newer
features like 'npm' are obviously not going to be supported on older
released images.

Sample invocation:

    # docker stop $(docker ps -a -q)
    # docker rmi $(docker images -a -q)

    # TEST_OPENSHIFT=true ./build.sh

Or, run each test itself directly:

    # IMAGE_NAME=registry.access.redhat.com/dotnet/dotnetcore-10-rhel7:latest OPENSHIFT_ONLY=true 1.0/test/run
    # IMAGE_NAME=registry.access.redhat.com/dotnet/dotnetcore-11-rhel7:latest OPENSHIFT_ONLY=true 1.1/test/run