Add DOTNET_ASSEMBLY_NAME envvar (#42)
It is used to select the assembly to run. This must NOT include the .dll
extension. Set this to the output assembly name specified in project.json
(specified via buildOptions/outputName or name) .

For project.json, the assembly name defaults to project.json's parent
folder if none of the above properties are specified.
The parent folder is used as the default value for DOTNET_ASSEMBLY_NAME.
When project.json is at the context directory, the parent folder name will
be 'src'. So, by default, this generates a 'src.dll' assembly. Setting
DOTNET_ASSEMBLY_NAME will cause the project to be built in a folder
named $DOTNET_ASSEMBLY_NAME which builds a $DOTNET_ASSEMBLY_NAME.dll.

Also note that sources in the built container will be in subfolder
$DOTNET_ASSEMBLY_NAME rather than in the current working directory
directly.