Add DOTNET_PUBLISH
Used to control whether the application should be build by executing dotnet build or dotnet publish. To publish the application set the value to Y. It is recommended to publish your application. For backwards compatibility, the default is ``. In the next major release of the builder, this variable will be removed and the builder will always publish the application.

Closes #61