Don't do an RID specific restore to avoid downloading runtime.linux-x64.microsoft.netcore.app
The output of the published folder can be reduced by doing an RID specific publish.
An RID specific publish requires an RID specific restore. The RID specific restore
drags in the microsoft.netcore.app linux-x64 package. This package weighs about 28MB
and is used for self-contained deployments.

So we won't optimize the size of the published folder at the cost of increasing the
restore bandwidth.