Print sdk version after fetching sources.
This makes the echoed version reflect the sdk that may be selected
via a global.json file in the source repository.

This commit also changes from using 'cp' to 'mv'. This avoids
having the sources stored twice in the image. This is similar to s2i node images.

An additional check is added to check if the '/tmp/src' directory exists.
Without that check the sdk version test would fail because it doesn't have
sources. Perhaps this can also happen in practice, in particular when setting using
DEV_MODE with an empty git repository.

For consistency, the changes are also applied to 1.0 and 1.1.