2.1: Add DOTNET_SSL_DIRS (#199)
Used to specify a list of folders with additional certificates to trust. The certificates are
trusted by each process that runs during the build and all processes that runs in the image
after the build (including the application that was built). The folders can be absolute paths
(starting with `/`) or paths in the source repository (e.g. `certificates`).