2.1, 2.2: Introduce DOTNET_DATA_DIR
Our images don't have an empty folder that is writable by the user and
could be used for persistent data.

This makes available a preferred location for such data.

The environment variable may be picked up by the application and used
as a location to store the data.

When doing a volume mount for persistance, the container platform creates
the folder in the container when it doesn't exist. By having the folder
already available in the image, we ensure it also exists when the image
is used in a non-persistent way.