2.1, 2.2: add support for incremental builds
For incremental builds, we save and restore the NuGet packages.

The user needs to set 'DOTNET_RM_NUGET' to 'false' in order to keep the NuGet
packages, so they can be re-used.

Note that the builder does not remove unused NuGet packages on incremental
builds. Over time the incremental application image may contain unused packages.
To clean up those packages, the user should occasionally perform a
non-incremental build or perform a build with `DOTNET_RM_NUGET` set to `true`.