2.1, 2.2: Improve incremental build speed (#265)
Incremental builds have a cost of fetching the previous image from
the registry.

This change reduces the number of artifacts we keep in the application
image for an incremental build. Instead of keeping the entire
~/.nuget/packages folder, we only keep the *.nupkg files. On the next
build, we pass those files to a separate restore step.

As a bonus, this change causes packages that are no longer used from
the previous build to no longer be kept around.