Add DOTNET_RESTORE_CONFIGFILE
This envvar can be used to specify a NuGet.Config file to be used for restore operations.
It allows for more advanced use-cases than the existing DOTNET_RESTORE_SOURCES such as
specifying credentials, or providing the configuration via a file in the repository,
or a volume that is mounted in the container.