Enable use of 'dotnet' in Dockerfile RUN instructions on scl-based images. (#330)
The 'dotnet' command cannot be used directly from Dockerfile RUN instructions
because the scl is not enabled.

This adds a script named 'dotnet' on PATH which enables the scl and invoke
'dotnet' in it. Once the scl is enabled, the 'dotnet' host takes preference
over the script because it is earlier on the PATH variable.