Get tests to pass on Fedora
Also, enable .NET 6 tests on Fedora.

The 3.1 tests were failing because the output of dotnet restore changed
in the 3.1.4xx series to remove a string that the tests were looking
for. Adjust that to match what we do in the 5.0 and 6.0 tests and
manually add that restore text in the output.

Fix up versions for packages so tests all pass by default.