Exit out on dotnet restore failure
Currently, dotnet restore is an if condition check, which may result in
the assemble scripts continuing to run even if the restore operation
fails. This is unexpected and not desired. Fix that to make the assemble
script fail if dotnet restore fails.

And fix the tests to rely on the messages printed out by dotnet restore,
instead of something shown by the assemble script.