Fix `null byte in input` warnings when running tests
When running tests, I see several warnings like these:

    6.0/build/test/run: line 41: warning: command substitution: ignored null byte in input

It looks like these are coming from the `readlink -zf ...` command. The
`-z` flag tells readlink to end the output with null.

That doesn't seem to be useful for us: we use a single file name as
input to readlink, so using null to separate multiple file names in the
output is of limited value.