Flag changes in packages shipped in containers (#442)
It's possible for customers to take on a dependency on any package
shipped in a container. We shouldn't break that. In other words, any
changes in packages installed in our containers should require a manual
review.

That review should be useful even between major versions and should get
a manual review as well.