8.0: remove nss_wrapper to reduce image size.
The nss_wrapper is used to introduce a known user for the random uid
that OpenShift uses while running a container.

With OpenShift 4, such a user will already automatically be
added in the /etc/passwd file without needing the nss_wrapper.

For OpenShift 3, removing the nss_wrapper leads to no user being
known. The only side-effect for .NET is that the Environment.UserName
will return an empty string.