8.0: remove Node.js.
Node.js was included since .NET Core 1.0 in our images because that
was the front-end framework used by the .NET templates at the time
and there was no .NET-based alternative available.

For .NET 8, the SDK no longer includes any Node.js-based templates,
and the templates use Blazor for the front-end.

In line with these upstream changes, we are removing Node.js from
the .NET 8 SDK image.

This is a breaking changes. Users migrating to the .NET 8 image that
depend on Node.js will need to adjust their build. One option is to
created an image derived from the .NET 8 image and install Node.js.
Another option is to change to use a multi-stage build which
builds the front-end in a separate stage, which may use the Node.js
UBI images.