Stop bash from printing locale warnings when invoked through an Exec Task.
MSBuild Exec Tasks set the locale to en_US. Because that locale is not
available in our images, it causes bash to print warnings to standard
error. This error output will even cause the Task to fail when it has
LogStandardErrorAsError set.

This installs the en_US locale to stop bash from printing these warnings.