9.0: remove binary s2i from runtime image.
This removes the s2i feature from the runtime image.
This feature enabled building an application image on top of
a runtime image by using a pre-built application as source
instead of source code.

This required a custom pipeline and we remove the feature
in favor of pipelines that uses the .NET SDK instead.
The SDK can build an application image on top of a runtime image
directly from source code.

Note that the SDK image also supports building an application
image using a pre-built application as source (though using the
SDK image as a base image). This can be used as a stop gap
while users adopt an SDK-based pipeline.