Do not run rpm --verify (#499)
It seems to have become a recommendation since:
https://github.com/sclorg/s2i-perl-container/pull/67. That was when yum
would report success even if some packages could not be installed. `rpm
-v` would act as a sanity check and confirm that the packages were
actually installed.

But modern versions of dnf and microdnf are much better about this and
error out when packages can't be installed. We don't need to `rpm -v` or
`rpm --verify` anymore.