Set up tests to be run via GitHub Actions
This allows us to shift-left and test the container images on every
change to this repo.

The test-matrix will need to be updated for every future version of
.NET.

This only tests things on Intel (x86_64). It would be nice to test on
other architectures (aarch64, ppc64le, s390x), but there doesn't seem to
be an obvious way to do that via GitHub actions. Perhaps we can find an
alternative service later and add that.

This does use a hardcoded version of s2i. It will need to be updated in
the future. I was trying to use the latest version, but the
currently-latest version doesn't have usable binaries avalabile as part
of the release:
https://github.com/openshift/source-to-image/releases/tag/v1.5.0

It's not obvious, but both podman and docker binaries are installed in
the GitHub runner. The build script prints the following to identify
what's going on:

    [INFO] Using podman

While running tests, I observed a numger of `broken pipe` errors. I
replaced `grep -q` with `grep >/dev/null` to make `grep` stop from
existing on first match, prematurely closing the pipe.