Support testing image names with a hash/tag
I am trying to use the test suite to test images with an image from a
(non-local) regitry and the image name contains a hash or a tag. It
would be great to support testing that directly without having to retag
the image.

The only real change is to adjust the name used for the s2i app.