Automatically infer runtime image when testing the build images
Instead of guessing/assuming the base runtime image, compute it from the
exact base (`FROM`) of the build container image. A nice advantage is
that if build image uses an exact image-with-hash as the base image, the
tests will run against the exact same image whenever they are re-run.

Also log the runtime image name for later diagnosis.