Update container labels for .NET 10.
We no longer set the 'version' label so that the label from the base image is preserved.

The 'dotnet_version' label is used for the version of the component.
For the SDK image, it is the full SDK version, like 10.0.100.
For the runtime image, it is the full ASP.NET Core runtime version, like 10.0.0.

The 'dotnet_version_major_minor' contains only the major.minor part, like 10.0.