.NET 10: further align the runtime image with Microsoft images.
SDK specific envvars are moved from the runtime to the SDK image.

All s2i related functionality is moved into the SDK image.

WORKDIR, CMD, and PATH envvar are set to match with Microsoft images.

The DOTNET_DATA_PATH dir is removed. It is left to the user's application to create such a directory (if needed).